backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
python-jose[cryptography]

